app/requirements.txt
Flask
psycopg2-binary==2.9.7
prometheus_client
werkzeug
